LayoutTests/imported/w3c/web-platform-tests/service-workers/stub-3.2-navigator-service-worker-expected.txt

CONSOLE MESSAGE: line 871: Invalid object member name RegistrationOptionList

Harness Error (FAIL), message = Invalid object member name RegistrationOptionList

FAIL Partial interface Navigator: original interface defined assert_true: Original interface should be defined expected true got false
PASS ServiceWorkerContainer interface: existence and properties of interface object 
PASS ServiceWorkerContainer interface object length 
PASS ServiceWorkerContainer interface object name 
PASS ServiceWorkerContainer interface: existence and properties of interface prototype object 
PASS ServiceWorkerContainer interface: existence and properties of interface prototype object's "constructor" property 
PASS ServiceWorkerContainer interface: existence and properties of interface prototype object's @@unscopables property 
PASS ServiceWorkerContainer interface: attribute ready 
FAIL ServiceWorkerContainer interface: operation getAll() assert_own_property: interface prototype object missing non-static operation expected property "getAll" missing
PASS ServiceWorkerContainer interface: operation register(DOMString, RegistrationOptionList) 
FAIL ServiceWorkerContainer interface: operation unregister(DOMString) assert_own_property: interface prototype object missing non-static operation expected property "unregister" missing
FAIL ServiceWorkerContainer interface: attribute onupdatefound assert_true: The prototype object must have a property "onupdatefound" expected true got false
PASS ServiceWorkerContainer interface: attribute oncontrollerchange 
FAIL ServiceWorkerContainer interface: attribute onreloadpage assert_true: The prototype object must have a property "onreloadpage" expected true got false
FAIL ServiceWorkerContainer interface: attribute onerror assert_true: The prototype object must have a property "onerror" expected true got false
FAIL ServiceWorkerContainer must be primary interface of throw new Error ('No object defined for the ServiceWorkerContainer interface') assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
FAIL Stringification of throw new Error ('No object defined for the ServiceWorkerContainer interface') assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
FAIL ServiceWorkerContainer interface: throw new Error ('No object defined for the ServiceWorkerContainer interface') must have own property "installing" assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
FAIL ServiceWorkerContainer interface: throw new Error ('No object defined for the ServiceWorkerContainer interface') must have own property "waiting" assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
FAIL ServiceWorkerContainer interface: throw new Error ('No object defined for the ServiceWorkerContainer interface') must have own property "active" assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
FAIL ServiceWorkerContainer interface: throw new Error ('No object defined for the ServiceWorkerContainer interface') must have own property "controller" assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
FAIL ServiceWorkerContainer interface: throw new Error ('No object defined for the ServiceWorkerContainer interface') must inherit property "ready" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
FAIL ServiceWorkerContainer interface: throw new Error ('No object defined for the ServiceWorkerContainer interface') must inherit property "getAll()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
FAIL ServiceWorkerContainer interface: throw new Error ('No object defined for the ServiceWorkerContainer interface') must inherit property "register(DOMString, RegistrationOptionList)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
FAIL ServiceWorkerContainer interface: calling register(DOMString, RegistrationOptionList) on throw new Error ('No object defined for the ServiceWorkerContainer interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
FAIL ServiceWorkerContainer interface: throw new Error ('No object defined for the ServiceWorkerContainer interface') must inherit property "unregister(DOMString)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
FAIL ServiceWorkerContainer interface: calling unregister(DOMString) on throw new Error ('No object defined for the ServiceWorkerContainer interface') with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
FAIL ServiceWorkerContainer interface: throw new Error ('No object defined for the ServiceWorkerContainer interface') must inherit property "onupdatefound" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
FAIL ServiceWorkerContainer interface: throw new Error ('No object defined for the ServiceWorkerContainer interface') must inherit property "oncontrollerchange" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
FAIL ServiceWorkerContainer interface: throw new Error ('No object defined for the ServiceWorkerContainer interface') must inherit property "onreloadpage" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
FAIL ServiceWorkerContainer interface: throw new Error ('No object defined for the ServiceWorkerContainer interface') must inherit property "onerror" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "Error: No object defined for the ServiceWorkerContainer interface"
